Move code related to bone spheres to separate handler
