+++ BoneSpheresHandler.h
@@ -1,0 +1,73 @@
+#pragma once
+
+#include "f4se/PapyrusEvents.h"
+
+namespace F4VRBody {
+
+	constexpr const char* BONE_SPHERE_EVEN_NAME = "OnBoneSphereEvent";
+
+	enum BoneSphereEvent {
+		BoneSphereEvent_None = 0,
+		BoneSphereEvent_Enter = 1,
+		BoneSphereEvent_Exit = 2,
+		BoneSphereEvent_Holster = 3,
+		BoneSphereEvent_Draw = 4
+	};
+
+	class BoneSphere {
+	public:
+		BoneSphere() {
+			radius = 0;
+			bone = nullptr;
+			stickyRight = false;
+			stickyLeft = false;
+			turnOnDebugSpheres = false;
+			offset.x = 0;
+			offset.y = 0;
+			offset.z = 0;
+			debugSphere = nullptr;
+		}
+
+		BoneSphere(float a_radius, NiNode* a_bone, NiPoint3 a_offset) : radius(a_radius), bone(a_bone), offset(a_offset) {
+			stickyRight = false;
+			stickyLeft = false;
+			turnOnDebugSpheres = false;
+			debugSphere = nullptr;
+		}
+
+		float radius;
+		NiNode* bone;
+		NiPoint3 offset;
+		bool stickyRight;
+		bool stickyLeft;
+		bool turnOnDebugSpheres;
+		NiNode* debugSphere;
+	};
+
+	class BoneSpheresHandler
+	{
+	public:
+		void onFrameUpdate();
+
+		void holsterWeapon();
+		void drawWeapon();
+
+		UInt32 registerBoneSphere(float radius, BSFixedString bone);
+		UInt32 registerBoneSphereOffset(float radius, BSFixedString bone, VMArray<float> pos);
+		void destroyBoneSphere(UInt32 handle);
+		void registerForBoneSphereEvents(VMObject* thisObject);
+		void unRegisterForBoneSphereEvents(VMObject* thisObject);
+		void toggleDebugBoneSpheres(bool turnOn);
+		void toggleDebugBoneSpheresAtBone(UInt32 handle, bool turnOn);
+	
+	private:
+		void detectBoneSphere();
+		void handleDebugBoneSpheres();
+
+		RegistrationSetHolder<NullParameters> _boneSphereEventRegs;
+
+		std::map<UInt32, BoneSphere*> _boneSphereRegisteredObjects;
+		UInt32 _nextBoneSphereHandle = 1;
+		UInt32 _curDevice = 0;
+	};
+}
