+++ Fallout4VR_Body.vcxproj
@@ -172,6 +172,7 @@ copy "$(TargetDir)$(TargetName).pdb" "%FRIK_MOD_PATH%" /Y</Command>
     </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="BoneSpheresHandler.cpp" />
     <ClCompile Include="BSFlattenedBoneTree.cpp" />
     <ClCompile Include="Config.cpp" />
     <ClCompile Include="ConfigurationMode.cpp" />
@@ -204,6 +205,7 @@ copy "$(TargetDir)$(TargetName).pdb" "%FRIK_MOD_PATH%" /Y</Command>
     <ClInclude Include="api\PapyrusVRTypes.h" />
     <ClInclude Include="api\VRHookAPI.h" />
     <ClInclude Include="api\VRManagerAPI.h" />
+    <ClInclude Include="BoneSpheresHandler.h" />
     <ClInclude Include="BSFlattenedBoneTree.h" />
     <ClInclude Include="Config.h" />
     <ClInclude Include="ConfigurationMode.h" />
