+++ Fallout4VR_Body.vcxproj.filters
@@ -84,6 +84,9 @@
     <ClCompile Include="CullGeometryHandler.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="BoneSpheresHandler.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="hook.h">
@@ -189,6 +192,9 @@
       <Filter>Include Files</Filter>
     </ClInclude>
     <ClInclude Include="CullGeometryHandler.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="BoneSpheresHandler.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
